ERR: Remove warnings on Visual Studio 7

+++ Source/CTest/Curl/base64.c
@@ -127,7 +127,7 @@ int Curl_base64_encode(const void *inp, int insize, char **outptr)
   char *indata = (char *)inp;
 
   if(0 == insize)
-    insize = strlen(indata);
+    insize = (int)strlen(indata);
 
   base64data = output = (char*)malloc(insize*4/3+4);
   if(NULL == output)
@@ -177,7 +177,7 @@ int Curl_base64_encode(const void *inp, int insize, char **outptr)
   *output=0;
   *outptr = base64data; /* make it return the actual data memory */
 
-  return strlen(base64data); /* return the length of the new data */
+  return (int)strlen(base64data); /* return the length of the new data */
 }
 /* ---- End of Base64 Encoding ---- */
